refactor: renames hooks index files
