+++ src/hooks/use-api/helpers/unit-measure.ts
@@ -1,7 +1,7 @@
 import { UnitMeasure } from '@src/store/slices/unit-measure/types';
 import * as services from '@src/services/app/unit-measure';
 import { translateGetError } from '@src/utils';
-import { OnError, Response } from './types';
+import { OnError, Response } from '../types';
 
 type HandleProduct = (onError: OnError) => {
   getAll: () => Promise<Response<UnitMeasure[]>>;
