+++ src/hooks/use-auth/use-sign-up/helpers/index.ts
@@ -1,0 +1,4 @@
+export { handleRegisterConsumer } from './register-consumer';
+export { handleRegisterProducer } from './register-producer';
+export { handleReserveCredentials } from './reserve-credentials';
+export { handleReservePhone } from './reserve-phone';
