+++ src/hooks/use-auth/use-sign-up/helpers/register-consumer.ts
