+++ src/hooks/use-auth/use-sign-up/helpers/reserve-credentials.ts
